move some helpers into core
