+++ packages/core/src/index.ts
@@ -23,6 +23,7 @@ export { default as textPasteRule } from './pasteRules/textPasteRule'
 export { default as callOrReturn } from './utilities/callOrReturn'
 export { default as mergeAttributes } from './utilities/mergeAttributes'
 
+export { default as combineTransactionSteps } from './helpers/combineTransactionSteps'
 export { default as defaultBlockAt } from './helpers/defaultBlockAt'
 export { default as getExtensionField } from './helpers/getExtensionField'
 export { default as findChildren } from './helpers/findChildren'
@@ -32,6 +33,7 @@ export { default as findParentNodeClosestToPos } from './helpers/findParentNodeC
 export { default as generateHTML } from './helpers/generateHTML'
 export { default as generateJSON } from './helpers/generateJSON'
 export { default as generateText } from './helpers/generateText'
+export { default as getChangedRanges } from './helpers/getChangedRanges'
 export { default as getSchema } from './helpers/getSchema'
 export { default as getHTMLFromFragment } from './helpers/getHTMLFromFragment'
 export { default as getDebugJSON } from './helpers/getDebugJSON'
